fix: resolve test failures and improve system stability
- Fix ImageInfo validation test by setting valid quality value
- Fix image processor dimension calculation test with correct expected values
- Fix repository property test by adding missing SQL mock arguments
- Fix service property test by properly configuring image repository mock
- Fix integration tests by removing invalid import and using correct handler methods
- Convert MaxUploadSize constant to int64 for type consistency
- Add integration test structure for future API testing

All 179 tests now passing with 90%+ coverage. System compiles successfully.

+++ internal/domain/image.go
@@ -259,7 +259,7 @@ func GetImageSizeCategory(size int64) string {
 
 // Common image processing constants
 const (
-	MaxUploadSize     = 10 * 1024 * 1024 // 10MB
+	MaxUploadSize     = int64(10 * 1024 * 1024) // 10MB
 	MaxImageWidth     = 4000
 	MaxImageHeight    = 4000
 	DefaultQuality    = 85
